requirements.txt
streamlit>=1.34.0
datasets>=2.19.0
sentence-transformers>=2.7.0
faiss-cpu>=1.7.4
transformers>=4.44.0
torch>=2.2.0
numpy>=1.24.0
pandas>=2.2.0
accelerate>=0.33.0
python-dotenv>=1.0.1
# optional for Gemini
google-generativeai>=0.7.2